Fixing error where rounding fails on integer (NTAW-553)

+++ app/ms1/task_functions.py
@@ -1154,10 +1154,12 @@ def occ_drop_df(df, docs, df_flagged, Mean_Samples):
     df["Final Occurrence Count"] = docs["Final Occurrence Count"]
     df_flagged["Final Occurrence Count (with flags)"] = docs["Final Occurrence Count (with flags)"]
     # Calculate Final Occurrence Percentage
-    df["Final Occurrence Percentage"] = (df["Final Occurrence Count"] / df["Possible Occurrence Count"]).round(2)
+    df["Final Occurrence Percentage"] = (
+        (df["Final Occurrence Count"] / df["Possible Occurrence Count"]).float().round(2)
+    )
     df_flagged["Final Occurrence Percentage (with flags)"] = (
-        df_flagged["Final Occurrence Count (with flags)"] / df_flagged["Possible Occurrence Count"]
-    ).round(2)
+        (df_flagged["Final Occurrence Count (with flags)"] / df_flagged["Possible Occurrence Count"]).float().round(2)
+    )
     return df, df_flagged
